requirements.txt
Flask
Flask-SocketIO
qrcode
pillow
pyinstaller
